+++ src/components/Sidebar/exitModal/index.jsx
@@ -1,0 +1,98 @@
+import axios from "axios";
+
+import { useNavigate } from "react-router-dom";
+import { useMutation } from "@tanstack/react-query";
+import { useHttp } from "../../../hook/useHttp";
+import { ToastContainer, toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
+import { MenuCloseIcons, UserExitIcon } from "../../../assets/icons";
+export default function ExitModal({ setModalOpen, modalOpen }) {
+  const { request } = useHttp()
+
+  const navigate = useNavigate();
+  // -----------------------Seller Delete---------------
+  const HandleLogOutSeller = useMutation(() => {
+    return request({ url: `/logout`, method: "POST", token: true });
+  });
+  const logOutHandle = () => {
+    HandleLogOutSeller.mutate({}, {
+      onSuccess: res => {
+        if (res?.message) {
+          localStorage.clear();
+          navigate("/login-seller")
+          toast.success(`${res?.message}`, {
+            position: "top-right",
+            autoClose: 3000,
+            hideProgressBar: false,
+            closeOnClick: true,
+            pauseOnHover: true,
+            draggable: true,
+            progress: undefined,
+            theme: "light",
+          });
+          // setDressInfo({ ...dressInfo, logOutSeller: false })
+        }
+      },
+      onError: err => {
+
+      }
+    })
+  }
+
+  return (
+    <div className={`w-full px-4 md:px-10 hidden md:block`}>
+      <ToastContainer
+        position="top-right"
+        autoClose={5000}
+        limit={4}
+        hideProgressBar={false}
+        newestOnTop={false}
+        closeOnClick
+        rtl={false}
+        pauseOnFocusLoss
+        draggable
+        pauseOnHover
+        theme="colored"
+      />
+      <div
+        className={`fixed cursor-pointer z-[200] inset-0 w-full h-full bg-black opacity-40 ${modalOpen ? "" : "hidden"
+          }`}
+        onClick={() => setModalOpen(false)}
+      ></div>
+      <section
+        className={`max-w-[90%] md:max-w-[550px] z-[201] mx-auto w-full flex-col h-fit bg-white fixed py-[30px] md:py-[35px] px-[20px] md:px-[50px] rounded-t-lg rounded-b-lg md:top-[50%] duration-300 overflow-hidden left-1/2 right-1/2 translate-x-[-50%] translate-y-[-50%] ${modalOpen ? "bottom-0 flex" : "hidden z-[-10]"
+          }`}
+      >
+        <button
+          onClick={() => setModalOpen(false)}
+          type="button"
+          className="absolute  right-3 top-3 w-5 h-5 ">
+          <MenuCloseIcons
+            className="w-full h-full"
+            colors={"#a1a1a1"} />
+        </button>
+        <div className="flex flex-col justify-center items-center gap-y-2 ll:gap-y-4">
+          <span className=" text-black text-lg xs:text-xl not-italic font-AeonikProMedium text-center">
+            Вы уверены?
+          </span>
+        </div>
+        <div className="w-full flex items-center justify-between mt-5 xs:mt-10 gap-x-2">
+          <button
+            onClick={() => setModalOpen(false)}
+            type="button"
+            className="w-1/2 xs:w-[45%] active:scale-95  active:opacity-70 flex items-center justify-center rounded-[12px] border border-textBlueColor text-textBlueColor bg-white h-[42px] px-4  text-center text-base not-italic font-AeonikProMedium">
+            Oтмена
+          </button>
+          <button
+            onClick={logOutHandle}
+            type="button"
+            className="w-1/2 xs:w-[45%] active:scale-95  active:opacity-70 flex items-center gap-x-2 justify-center rounded-[12px] border border-textRedColor text-white bg-[#FF4747]  h-[42px] px-4  text-center text-base not-italic font-AeonikProMedium">
+            <UserExitIcon colors={"#fff"} />{" "}
+            <span>Выйти</span>
+          </button>
+        </div>
+
+      </section >
+    </div >
+  );
+}
